Add a base test class
